python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.28
pandas==2.1.3
numpy==1.25.2
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
pydantic==2.5.0
python-multipart==0.0.6

